+++ src/templates/20-million.js
@@ -15,7 +15,7 @@ import OurPartners from "../components/OurPartners";
 import ScholarshipProjects from "../components/ScholarshipProjects";
 import ScholarshipSuccessCases from "../components/ScholarshipSuccessCases";
 import BenefitsAndCharts from "../components/BenefitsAndCharts";
-import { TwoColumn } from "../components/Landing";
+import TwoColumn from "../components/TwoColumn/index.js";
 
 const TwentyMillion = ({ data, pageContext, yml }) => {
   const { session } = React.useContext(SessionContext);
