+++ src/templates/job-guarantee.js
@@ -22,7 +22,7 @@ import LeadForm from "../components/LeadForm";
 import { Circle } from "../components/BackgroundDrawing";
 import Icon from "../components/Icon";
 import ScholarshipSuccessCases from "../components/ScholarshipSuccessCases";
-import { TwoColumn } from "../components/Landing";
+import TwoColumn from "../components/TwoColumn/index.js";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
